pyproject.toml
[project]
name = "madera-mcp"
version = "0.1.0"
description = "MCP server for PDF triage hints - AI toolbox for document processing"
requires-python = ">=3.12"
readme = "README.md"
authors = [
    { name = "MADERA Team" }
]

dependencies = [
    # MCP & API
    "mcp>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.4",

    # Database
    "sqlalchemy>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",

    # Task Queue
    "celery>=5.4.0",
    "redis>=5.0.0",

    # Settings & Config
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",

    # PDF/Image Processing
    "pypdf>=4.3.0",
    "pdf2image>=1.17.0",
    "pillow>=10.4.0",
    "opencv-python>=4.10.0",
    "pytesseract>=0.3.10",

    # AI SDKs
    "google-generativeai>=0.8.0",
    "anthropic>=0.39.0",
    "openai>=1.54.0",

    # Storage
    "boto3>=1.34.0",
    "httpx>=0.27.0",

    # Utilities
    "phonenumbers>=8.13.0",
    "langdetect>=1.0.9",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "black>=24.0.0",
    "ruff>=0.6.0",
    "ipython>=8.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["madera"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
